refactor(sales): split sales_writer into modules
